♻️(PostExperience.svelte): get data from PostExperienceQuestions
